refactor: move formatted_error under src
